handle deployments
