+++ exporter/chqk8sentitygraphexporter/go.mod
@@ -156,4 +156,5 @@ require (
 replace (
 	github.com/cardinalhq/cardinalhq-otel-collector/extension/chqconfigextension v0.0.0 => ../../extension/chqconfigextension
 	github.com/cardinalhq/cardinalhq-otel-collector/internal v0.0.0 => ../../internal
+	github.com/cardinalhq/oteltools => ../../../oteltools
 )
