+++ exporter/chqk8sentitygraphexporter/internal/objecthandler/handler.go
@@ -20,6 +20,7 @@ import (
 	"go.opentelemetry.io/collector/pdata/pcommon"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 
+	convertappsv1 "github.com/cardinalhq/cardinalhq-otel-collector/exporter/chqk8sentitygraphexporter/internal/objecthandler/apps/v1"
 	"github.com/cardinalhq/cardinalhq-otel-collector/exporter/chqk8sentitygraphexporter/internal/objecthandler/baseobj"
 	"github.com/cardinalhq/cardinalhq-otel-collector/exporter/chqk8sentitygraphexporter/internal/objecthandler/converterconfig"
 	convertv1 "github.com/cardinalhq/cardinalhq-otel-collector/exporter/chqk8sentitygraphexporter/internal/objecthandler/v1"
@@ -59,6 +60,7 @@ func (h *converterImpl) installConverters() {
 	h.converters[objectSelector{"v1", "Pod"}] = convertv1.ConvertPod
 	h.converters[objectSelector{"v1", "ConfigMap"}] = convertv1.ConvertConfigMap
 	h.converters[objectSelector{"v1", "Secret"}] = convertv1.ConvertSecret
+	h.converters[objectSelector{"apps/v1", "Deployment"}] = convertappsv1.ConvertDeployment
 }
 
 // Feed takes a set of attributes and an object and converts the object to a PackagedObject.
