+++ exporter/chqk8sentitygraphexporter/internal/objecthandler/v1/pod_test.go
@@ -21,11 +21,9 @@ import (
 	"github.com/cardinalhq/cardinalhq-otel-collector/exporter/chqk8sentitygraphexporter/internal/objecthandler/baseobj"
 	"github.com/cardinalhq/cardinalhq-otel-collector/exporter/chqk8sentitygraphexporter/internal/objecthandler/converterconfig"
 	"github.com/cardinalhq/oteltools/pkg/graph/graphpb"
-	mapset "github.com/deckarep/golang-set/v2"
 	"github.com/stretchr/testify/assert"
 	"google.golang.org/protobuf/types/known/timestamppb"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 )
 
@@ -125,589 +123,6 @@ func TestPodIPs(t *testing.T) {
 	}
 }
 
-func TestConvertPodResources(t *testing.T) {
-	tests := []struct {
-		name     string
-		input    corev1.ResourceRequirements
-		expected map[string]string
-	}{
-		{
-			name: "No resources",
-			input: corev1.ResourceRequirements{
-				Requests: corev1.ResourceList{},
-				Limits:   corev1.ResourceList{},
-			},
-			expected: nil,
-		},
-		{
-			name: "Only requests",
-			input: corev1.ResourceRequirements{
-				Requests: corev1.ResourceList{
-					corev1.ResourceCPU:    resource.MustParse("500m"),
-					corev1.ResourceMemory: resource.MustParse("128Mi"),
-				},
-				Limits: corev1.ResourceList{},
-			},
-			expected: map[string]string{
-				"requests.cpu":    "500m",
-				"requests.memory": "128Mi",
-			},
-		},
-		{
-			name: "Only limits",
-			input: corev1.ResourceRequirements{
-				Requests: corev1.ResourceList{},
-				Limits: corev1.ResourceList{
-					corev1.ResourceCPU:    resource.MustParse("1"),
-					corev1.ResourceMemory: resource.MustParse("256Mi"),
-				},
-			},
-			expected: map[string]string{
-				"limits.cpu":    "1",
-				"limits.memory": "256Mi",
-			},
-		},
-		{
-			name: "Both requests and limits",
-			input: corev1.ResourceRequirements{
-				Requests: corev1.ResourceList{
-					corev1.ResourceCPU:    resource.MustParse("250m"),
-					corev1.ResourceMemory: resource.MustParse("64Mi"),
-				},
-				Limits: corev1.ResourceList{
-					corev1.ResourceCPU:    resource.MustParse("500m"),
-					corev1.ResourceMemory: resource.MustParse("128Mi"),
-				},
-			},
-			expected: map[string]string{
-				"requests.cpu":    "250m",
-				"requests.memory": "64Mi",
-				"limits.cpu":      "500m",
-				"limits.memory":   "128Mi",
-			},
-		},
-		{
-			name: "Zero values are ignored",
-			input: corev1.ResourceRequirements{
-				Requests: corev1.ResourceList{
-					corev1.ResourceCPU:    resource.MustParse("0"),
-					corev1.ResourceMemory: resource.MustParse("0"),
-				},
-				Limits: corev1.ResourceList{
-					corev1.ResourceCPU:    resource.MustParse("0"),
-					corev1.ResourceMemory: resource.MustParse("0"),
-				},
-			},
-			expected: nil,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			result := convertPodResources(tt.input)
-			assert.Equal(t, tt.expected, result)
-		})
-	}
-}
-
-func TestContainerSecretNamesFromEnv(t *testing.T) {
-	tests := []struct {
-		name      string
-		container corev1.Container
-		expected  []string
-	}{
-		{
-			name: "No secrets in env",
-			container: corev1.Container{
-				Env:     []corev1.EnvVar{},
-				EnvFrom: []corev1.EnvFromSource{},
-			},
-			expected: []string{},
-		},
-		{
-			name: "Secrets in Env",
-			container: corev1.Container{
-				Env: []corev1.EnvVar{
-					{
-						Name: "SECRET_ENV",
-						ValueFrom: &corev1.EnvVarSource{
-							SecretKeyRef: &corev1.SecretKeySelector{
-								LocalObjectReference: corev1.LocalObjectReference{Name: "my-secret"},
-							},
-						},
-					},
-				},
-			},
-			expected: []string{"my-secret"},
-		},
-		{
-			name: "Secrets in EnvFrom",
-			container: corev1.Container{
-				EnvFrom: []corev1.EnvFromSource{
-					{
-						SecretRef: &corev1.SecretEnvSource{
-							LocalObjectReference: corev1.LocalObjectReference{Name: "my-envfrom-secret"},
-						},
-					},
-				},
-			},
-			expected: []string{"my-envfrom-secret"},
-		},
-		{
-			name: "Secrets in both Env and EnvFrom",
-			container: corev1.Container{
-				Env: []corev1.EnvVar{
-					{
-						Name: "SECRET_ENV",
-						ValueFrom: &corev1.EnvVarSource{
-							SecretKeyRef: &corev1.SecretKeySelector{
-								LocalObjectReference: corev1.LocalObjectReference{Name: "my-secret"},
-							},
-						},
-					},
-				},
-				EnvFrom: []corev1.EnvFromSource{
-					{
-						SecretRef: &corev1.SecretEnvSource{
-							LocalObjectReference: corev1.LocalObjectReference{Name: "my-envfrom-secret"},
-						},
-					},
-				},
-			},
-			expected: []string{"my-secret", "my-envfrom-secret"},
-		},
-		{
-			name: "Duplicate secrets are deduplicated",
-			container: corev1.Container{
-				Env: []corev1.EnvVar{
-					{
-						Name: "SECRET_ENV",
-						ValueFrom: &corev1.EnvVarSource{
-							SecretKeyRef: &corev1.SecretKeySelector{
-								LocalObjectReference: corev1.LocalObjectReference{Name: "duplicate-secret"},
-							},
-						},
-					},
-				},
-				EnvFrom: []corev1.EnvFromSource{
-					{
-						SecretRef: &corev1.SecretEnvSource{
-							LocalObjectReference: corev1.LocalObjectReference{Name: "duplicate-secret"},
-						},
-					},
-				},
-			},
-			expected: []string{"duplicate-secret"},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			seen := mapset.NewSet[string]()
-			containerSecretNamesFromEnv(seen, tt.container)
-			assert.ElementsMatch(t, tt.expected, seen.ToSlice())
-		})
-	}
-}
-
-func TestContainerSecretNamesFromVolumes(t *testing.T) {
-	tests := []struct {
-		name       string
-		podVolumes []corev1.Volume
-		container  corev1.Container
-		expected   []string
-	}{
-		{
-			name:       "No secrets in volumes",
-			podVolumes: []corev1.Volume{},
-			container: corev1.Container{
-				VolumeMounts: []corev1.VolumeMount{},
-			},
-			expected: []string{},
-		},
-		{
-			name: "Secret in volume",
-			podVolumes: []corev1.Volume{
-				{
-					Name: "my-volume",
-					VolumeSource: corev1.VolumeSource{
-						Secret: &corev1.SecretVolumeSource{
-							SecretName: "my-secret",
-						},
-					},
-				},
-			},
-			container: corev1.Container{
-				VolumeMounts: []corev1.VolumeMount{
-					{Name: "my-volume"},
-				},
-			},
-			expected: []string{"my-secret"},
-		},
-		{
-			name: "Multiple secrets in volumes",
-			podVolumes: []corev1.Volume{
-				{
-					Name: "volume-1",
-					VolumeSource: corev1.VolumeSource{
-						Secret: &corev1.SecretVolumeSource{
-							SecretName: "secret-1",
-						},
-					},
-				},
-				{
-					Name: "volume-2",
-					VolumeSource: corev1.VolumeSource{
-						Secret: &corev1.SecretVolumeSource{
-							SecretName: "secret-2",
-						},
-					},
-				},
-			},
-			container: corev1.Container{
-				VolumeMounts: []corev1.VolumeMount{
-					{Name: "volume-1"},
-					{Name: "volume-2"},
-				},
-			},
-			expected: []string{"secret-1", "secret-2"},
-		},
-		{
-			name: "Ignore default-token secret",
-			podVolumes: []corev1.Volume{
-				{
-					Name: "default-token-volume",
-					VolumeSource: corev1.VolumeSource{
-						Secret: &corev1.SecretVolumeSource{
-							SecretName: "default-token-abc123",
-						},
-					},
-				},
-			},
-			container: corev1.Container{
-				VolumeMounts: []corev1.VolumeMount{
-					{Name: "default-token-volume"},
-				},
-			},
-			expected: []string{},
-		},
-		{
-			name: "Mixed valid and default-token secrets",
-			podVolumes: []corev1.Volume{
-				{
-					Name: "valid-volume",
-					VolumeSource: corev1.VolumeSource{
-						Secret: &corev1.SecretVolumeSource{
-							SecretName: "valid-secret",
-						},
-					},
-				},
-				{
-					Name: "default-token-volume",
-					VolumeSource: corev1.VolumeSource{
-						Secret: &corev1.SecretVolumeSource{
-							SecretName: "default-token-abc123",
-						},
-					},
-				},
-			},
-			container: corev1.Container{
-				VolumeMounts: []corev1.VolumeMount{
-					{Name: "valid-volume"},
-					{Name: "default-token-volume"},
-				},
-			},
-			expected: []string{"valid-secret"},
-		},
-		{
-			name: "Duplicate secrets are deduplicated",
-			podVolumes: []corev1.Volume{
-				{
-					Name: "volume-1",
-					VolumeSource: corev1.VolumeSource{
-						Secret: &corev1.SecretVolumeSource{
-							SecretName: "duplicate-secret",
-						},
-					},
-				},
-				{
-					Name: "volume-2",
-					VolumeSource: corev1.VolumeSource{
-						Secret: &corev1.SecretVolumeSource{
-							SecretName: "duplicate-secret",
-						},
-					},
-				},
-			},
-			container: corev1.Container{
-				VolumeMounts: []corev1.VolumeMount{
-					{Name: "volume-1"},
-					{Name: "volume-2"},
-				},
-			},
-			expected: []string{"duplicate-secret"},
-		},
-	}
-
-	conf := converterconfig.New("foo")
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			seen := mapset.NewSet[string]()
-			containerSecretNamesFromVolumes(conf, seen, tt.podVolumes, tt.container)
-			assert.ElementsMatch(t, tt.expected, seen.ToSlice())
-		})
-	}
-}
-
-func TestContainerConfigMapNamesFromVolumes(t *testing.T) {
-	tests := []struct {
-		name       string
-		podVolumes []corev1.Volume
-		container  corev1.Container
-		expected   []string
-	}{
-		{
-			name:       "No config maps in volumes",
-			podVolumes: []corev1.Volume{},
-			container: corev1.Container{
-				VolumeMounts: []corev1.VolumeMount{},
-			},
-			expected: []string{},
-		},
-		{
-			name: "Config map in volume",
-			podVolumes: []corev1.Volume{
-				{
-					Name: "my-volume",
-					VolumeSource: corev1.VolumeSource{
-						ConfigMap: &corev1.ConfigMapVolumeSource{
-							LocalObjectReference: corev1.LocalObjectReference{Name: "my-configmap"},
-						},
-					},
-				},
-			},
-			container: corev1.Container{
-				VolumeMounts: []corev1.VolumeMount{
-					{Name: "my-volume"},
-				},
-			},
-			expected: []string{"my-configmap"},
-		},
-		{
-			name: "Multiple config maps in volumes",
-			podVolumes: []corev1.Volume{
-				{
-					Name: "volume-1",
-					VolumeSource: corev1.VolumeSource{
-						ConfigMap: &corev1.ConfigMapVolumeSource{
-							LocalObjectReference: corev1.LocalObjectReference{Name: "configmap-1"},
-						},
-					},
-				},
-				{
-					Name: "volume-2",
-					VolumeSource: corev1.VolumeSource{
-						ConfigMap: &corev1.ConfigMapVolumeSource{
-							LocalObjectReference: corev1.LocalObjectReference{Name: "configmap-2"},
-						},
-					},
-				},
-			},
-			container: corev1.Container{
-				VolumeMounts: []corev1.VolumeMount{
-					{Name: "volume-1"},
-					{Name: "volume-2"},
-				},
-			},
-			expected: []string{"configmap-1", "configmap-2"},
-		},
-		{
-			name: "Ignore kube-root-ca.crt config map",
-			podVolumes: []corev1.Volume{
-				{
-					Name: "ca-volume",
-					VolumeSource: corev1.VolumeSource{
-						ConfigMap: &corev1.ConfigMapVolumeSource{
-							LocalObjectReference: corev1.LocalObjectReference{Name: "kube-root-ca.crt"},
-						},
-					},
-				},
-			},
-			container: corev1.Container{
-				VolumeMounts: []corev1.VolumeMount{
-					{Name: "ca-volume"},
-				},
-			},
-			expected: []string{},
-		},
-		{
-			name: "Mixed valid and kube-root-ca.crt config maps",
-			podVolumes: []corev1.Volume{
-				{
-					Name: "valid-volume",
-					VolumeSource: corev1.VolumeSource{
-						ConfigMap: &corev1.ConfigMapVolumeSource{
-							LocalObjectReference: corev1.LocalObjectReference{Name: "valid-configmap"},
-						},
-					},
-				},
-				{
-					Name: "ca-volume",
-					VolumeSource: corev1.VolumeSource{
-						ConfigMap: &corev1.ConfigMapVolumeSource{
-							LocalObjectReference: corev1.LocalObjectReference{Name: "kube-root-ca.crt"},
-						},
-					},
-				},
-			},
-			container: corev1.Container{
-				VolumeMounts: []corev1.VolumeMount{
-					{Name: "valid-volume"},
-					{Name: "ca-volume"},
-				},
-			},
-			expected: []string{"valid-configmap"},
-		},
-		{
-			name: "Duplicate config maps are deduplicated",
-			podVolumes: []corev1.Volume{
-				{
-					Name: "volume-1",
-					VolumeSource: corev1.VolumeSource{
-						ConfigMap: &corev1.ConfigMapVolumeSource{
-							LocalObjectReference: corev1.LocalObjectReference{Name: "duplicate-configmap"},
-						},
-					},
-				},
-				{
-					Name: "volume-2",
-					VolumeSource: corev1.VolumeSource{
-						ConfigMap: &corev1.ConfigMapVolumeSource{
-							LocalObjectReference: corev1.LocalObjectReference{Name: "duplicate-configmap"},
-						},
-					},
-				},
-			},
-			container: corev1.Container{
-				VolumeMounts: []corev1.VolumeMount{
-					{Name: "volume-1"},
-					{Name: "volume-2"},
-				},
-			},
-			expected: []string{"duplicate-configmap"},
-		},
-	}
-
-	conf := converterconfig.New("foo")
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			seen := mapset.NewSet[string]()
-			containerConfigMapNamesFromVolumes(conf, seen, tt.podVolumes, tt.container)
-			assert.ElementsMatch(t, tt.expected, seen.ToSlice())
-		})
-	}
-}
-
-func TestContainerConfigMapNamesFromEnv(t *testing.T) {
-	tests := []struct {
-		name      string
-		container corev1.Container
-		expected  []string
-	}{
-		{
-			name: "No config maps in env",
-			container: corev1.Container{
-				Env:     []corev1.EnvVar{},
-				EnvFrom: []corev1.EnvFromSource{},
-			},
-			expected: []string{},
-		},
-		{
-			name: "Config map in Env",
-			container: corev1.Container{
-				Env: []corev1.EnvVar{
-					{
-						Name: "CONFIG_ENV",
-						ValueFrom: &corev1.EnvVarSource{
-							ConfigMapKeyRef: &corev1.ConfigMapKeySelector{
-								LocalObjectReference: corev1.LocalObjectReference{Name: "my-configmap"},
-							},
-						},
-					},
-				},
-			},
-			expected: []string{"my-configmap"},
-		},
-		{
-			name: "Config map in EnvFrom",
-			container: corev1.Container{
-				EnvFrom: []corev1.EnvFromSource{
-					{
-						ConfigMapRef: &corev1.ConfigMapEnvSource{
-							LocalObjectReference: corev1.LocalObjectReference{Name: "my-envfrom-configmap"},
-						},
-					},
-				},
-			},
-			expected: []string{"my-envfrom-configmap"},
-		},
-		{
-			name: "Config maps in both Env and EnvFrom",
-			container: corev1.Container{
-				Env: []corev1.EnvVar{
-					{
-						Name: "CONFIG_ENV",
-						ValueFrom: &corev1.EnvVarSource{
-							ConfigMapKeyRef: &corev1.ConfigMapKeySelector{
-								LocalObjectReference: corev1.LocalObjectReference{Name: "my-configmap"},
-							},
-						},
-					},
-				},
-				EnvFrom: []corev1.EnvFromSource{
-					{
-						ConfigMapRef: &corev1.ConfigMapEnvSource{
-							LocalObjectReference: corev1.LocalObjectReference{Name: "my-envfrom-configmap"},
-						},
-					},
-				},
-			},
-			expected: []string{"my-configmap", "my-envfrom-configmap"},
-		},
-		{
-			name: "Duplicate config maps are deduplicated",
-			container: corev1.Container{
-				Env: []corev1.EnvVar{
-					{
-						Name: "CONFIG_ENV",
-						ValueFrom: &corev1.EnvVarSource{
-							ConfigMapKeyRef: &corev1.ConfigMapKeySelector{
-								LocalObjectReference: corev1.LocalObjectReference{Name: "duplicate-configmap"},
-							},
-						},
-					},
-				},
-				EnvFrom: []corev1.EnvFromSource{
-					{
-						ConfigMapRef: &corev1.ConfigMapEnvSource{
-							LocalObjectReference: corev1.LocalObjectReference{Name: "duplicate-configmap"},
-						},
-					},
-				},
-			},
-			expected: []string{"duplicate-configmap"},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			seen := mapset.NewSet[string]()
-			containerConfigMapNamesFromEnv(seen, tt.container)
-			assert.ElementsMatch(t, tt.expected, seen.ToSlice())
-		})
-	}
-}
-
 func TestConvertPod(t *testing.T) {
 	tests := []struct {
 		name     string
@@ -744,6 +159,12 @@ func TestConvertPod(t *testing.T) {
 						"podIPs": []any{
 							map[string]any{"ip": "10.0.0.1"},
 						},
+						"containerStatuses": []map[string]any{
+							{
+								"name":  "test-container",
+								"image": "test-image",
+							},
+						},
 					},
 				},
 			},
@@ -752,21 +173,31 @@ func TestConvertPod(t *testing.T) {
 					Name:      "test-pod",
 					Namespace: "default",
 				},
-				Phase:              "Running",
-				PhaseMessage:       "Pod is running",
-				ServiceAccountName: "test-service-account",
-				StartedAt: func() *timestamppb.Timestamp {
-					t := time.Date(2023, 1, 1, 0, 0, 0, 0, time.UTC)
-					return timestamppb.New(t)
-				}(),
-				Containers: []*graphpb.PodContainerSummary{
-					{
-						Name:  "test-container",
-						Image: &graphpb.ImageSummary{Image: "test-image"},
+				Spec: &graphpb.PodSpec{
+					ServiceAccountName: "test-service-account",
+					Containers: []*graphpb.PodContainerSpec{
+						{
+							Name:  "test-container",
+							Image: "test-image",
+						},
 					},
 				},
-				HostIps: []string{"192.168.1.1"},
-				PodIps:  []string{"10.0.0.1"},
+				Status: &graphpb.PodStatus{
+					Phase:        "Running",
+					PhaseMessage: "Pod is running",
+					StartedAt: func() *timestamppb.Timestamp {
+						t := time.Date(2023, 1, 1, 0, 0, 0, 0, time.UTC)
+						return timestamppb.New(t)
+					}(),
+					ContainerStatus: []*graphpb.PodContainerStatus{
+						{
+							Name:  "test-container",
+							Image: &graphpb.ImageSummary{Image: "test-image"},
+						},
+					},
+					HostIps: []string{"192.168.1.1"},
+					PodIps:  []string{"10.0.0.1"},
+				},
 			},
 			wantErr: false,
 		},
@@ -790,15 +221,31 @@ func TestConvertPod(t *testing.T) {
 					},
 					"status": map[string]any{
 						"phase": "Pending",
+						"containerStatuses": []map[string]any{
+							{
+								"name":  "test-container",
+								"image": "test-image",
+							},
+						},
 					},
 				},
 			},
 			expected: &graphpb.PodSummary{
-				Phase: "Pending",
-				Containers: []*graphpb.PodContainerSummary{
-					{
-						Name:  "test-container",
-						Image: &graphpb.ImageSummary{Image: "test-image"},
+				Spec: &graphpb.PodSpec{
+					Containers: []*graphpb.PodContainerSpec{
+						{
+							Name:  "test-container",
+							Image: "test-image",
+						},
+					},
+				},
+				Status: &graphpb.PodStatus{
+					Phase: "Pending",
+					ContainerStatus: []*graphpb.PodContainerStatus{
+						{
+							Name:  "test-container",
+							Image: &graphpb.ImageSummary{Image: "test-image"},
+						},
 					},
 				},
 			},
@@ -825,139 +272,9 @@ func TestConvertPod(t *testing.T) {
 				assert.Error(t, err)
 			} else {
 				assert.NoError(t, err)
-				tt.expected.BaseObject = baseobj.BaseFromUnstructured(cconf, tt.input)
+				tt.expected.BaseObject = baseobj.Make(cconf, &tt.input, tt.input.GetAPIVersion(), tt.input.GetKind())
 				assert.Equal(t, tt.expected, result)
 			}
-		})
-	}
-}
-
-func TestSetupContainers(t *testing.T) {
-	tests := []struct {
-		name     string
-		pod      corev1.Pod
-		expected []*graphpb.PodContainerSummary
-	}{
-		{
-			name: "Single container with resources, config maps, and secrets",
-			pod: corev1.Pod{
-				Spec: corev1.PodSpec{
-					Containers: []corev1.Container{
-						{
-							Name:  "test-container",
-							Image: "test-image",
-							Resources: corev1.ResourceRequirements{
-								Requests: corev1.ResourceList{
-									corev1.ResourceCPU:    resource.MustParse("500m"),
-									corev1.ResourceMemory: resource.MustParse("128Mi"),
-								},
-								Limits: corev1.ResourceList{
-									corev1.ResourceCPU:    resource.MustParse("1"),
-									corev1.ResourceMemory: resource.MustParse("256Mi"),
-								},
-							},
-							Env: []corev1.EnvVar{
-								{
-									Name: "CONFIG_ENV",
-									ValueFrom: &corev1.EnvVarSource{
-										ConfigMapKeyRef: &corev1.ConfigMapKeySelector{
-											LocalObjectReference: corev1.LocalObjectReference{Name: "my-configmap"},
-										},
-									},
-								},
-							},
-							EnvFrom: []corev1.EnvFromSource{
-								{
-									SecretRef: &corev1.SecretEnvSource{
-										LocalObjectReference: corev1.LocalObjectReference{Name: "my-secret"},
-									},
-								},
-							},
-							VolumeMounts: []corev1.VolumeMount{
-								{Name: "config-volume"},
-								{Name: "secret-volume"},
-							},
-						},
-					},
-					Volumes: []corev1.Volume{
-						{
-							Name: "config-volume",
-							VolumeSource: corev1.VolumeSource{
-								ConfigMap: &corev1.ConfigMapVolumeSource{
-									LocalObjectReference: corev1.LocalObjectReference{Name: "my-configmap"},
-								},
-							},
-						},
-						{
-							Name: "secret-volume",
-							VolumeSource: corev1.VolumeSource{
-								Secret: &corev1.SecretVolumeSource{
-									SecretName: "my-secret",
-								},
-							},
-						},
-					},
-				},
-			},
-			expected: []*graphpb.PodContainerSummary{
-				{
-					Name:  "test-container",
-					Image: &graphpb.ImageSummary{Image: "test-image"},
-					Resources: map[string]string{
-						"requests.cpu":    "500m",
-						"requests.memory": "128Mi",
-						"limits.cpu":      "1",
-						"limits.memory":   "256Mi",
-					},
-					ConfigMapNames: []string{"my-configmap"},
-					SecretNames:    []string{"my-secret"},
-				},
-			},
-		},
-		{
-			name: "Multiple containers sorted by name",
-			pod: corev1.Pod{
-				Spec: corev1.PodSpec{
-					Containers: []corev1.Container{
-						{
-							Name:  "b-container",
-							Image: "b-image",
-						},
-						{
-							Name:  "a-container",
-							Image: "a-image",
-						},
-					},
-				},
-			},
-			expected: []*graphpb.PodContainerSummary{
-				{
-					Name:  "a-container",
-					Image: &graphpb.ImageSummary{Image: "a-image"},
-				},
-				{
-					Name:  "b-container",
-					Image: &graphpb.ImageSummary{Image: "b-image"},
-				},
-			},
-		},
-		{
-			name: "No containers",
-			pod: corev1.Pod{
-				Spec: corev1.PodSpec{
-					Containers: []corev1.Container{},
-				},
-			},
-			expected: nil,
-		},
-	}
-
-	conf := &converterconfig.Config{}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			podSummary := &graphpb.PodSummary{}
-			setupContainers(conf, tt.pod, podSummary)
-			assert.Equal(t, tt.expected, podSummary.Containers)
 		})
 	}
 }
